Tests/LibWeb/Text/expected/HTML/document-open-cancels-delayed-parserless-completion.txt

Ready state after stale parserless completion: complete
Replacement load events: 1
Replacement pageshow events: 1
